LP-457 Move Jurisdiction enum class and fix imports
