+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/Jurisdiction.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.dto;
+package uk.gov.companieshouse.limitedpartnershipsapi.model;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonEnumDefaultValue;
