+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/dto/LimitedPartnershipPatchDto.java
@@ -3,6 +3,7 @@ package uk.gov.companieshouse.limitedpartnershipsapi.model.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Size;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.Jurisdiction;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.PartnershipNameEnding;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.PartnershipType;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.NameSize;
