+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/dto/validator/JurisdictionValidator.java
@@ -2,11 +2,12 @@ package uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
+import uk.gov.companieshouse.limitedpartnershipsapi.model.Jurisdiction;
 
-import static uk.gov.companieshouse.limitedpartnershipsapi.model.dto.Jurisdiction.UNKNOWN;
+import static uk.gov.companieshouse.limitedpartnershipsapi.model.Jurisdiction.UNKNOWN;
 
-public class JurisdictionValidator implements ConstraintValidator<ValidJurisdiction, uk.gov.companieshouse.limitedpartnershipsapi.model.dto.Jurisdiction> {
-    public boolean isValid(uk.gov.companieshouse.limitedpartnershipsapi.model.dto.Jurisdiction jurisdiction, ConstraintValidatorContext context) {
+public class JurisdictionValidator implements ConstraintValidator<ValidJurisdiction, Jurisdiction> {
+    public boolean isValid(Jurisdiction jurisdiction, ConstraintValidatorContext context) {
         if (jurisdiction == null) {
             return true;
         }
